Start applications multithreaded
With the new mod mail helper the program should run multithreaded
so it doesn't mix up features
