+++ conf/config.ini.dist
@@ -7,6 +7,7 @@ SUBREDDIT =
 REMOVE_SUBMISSIONS =
 MINIMUM_ACCOUNT_AGE =
 MINIMUM_COMMENT_KARMA =
+USER_AGENT = "BlotterTrax /r/listentothis submission bot"
 
 [YOUTUBE]
 KEY =
